+++ ui/components/canvas2d/src/AnnotationToolbar.svelte
@@ -18,8 +18,8 @@
   import type { Tool } from "./tools";
 
   // Exports
-  export let tools_lists: Array<Array<Tool>> = [];
-  export let selectedTool: Tool | null = null;
+  export let tools_lists: Array<Array<Tool>>;
+  export let selectedTool: Tool;
 
   function selectTool(tool: Tool) {
     // Prevent re-selecting the active tool
@@ -35,9 +35,9 @@
     >
       <div class="flex flex-col items-center">
         {#each tools as tool}
-          {#if tool["modes"]}
+          {#if tool.modes}
             <div class="relative group">
-              <button class="p-4" on:click={() => selectTool(tool["modes"][0])}>
+              <button class="p-4" on:click={() => selectTool(Object.values(tool.modes)[0])}>
                 <svg
                   xmlns="http://www.w3.org/2000/svg"
                   height="48"
@@ -45,7 +45,9 @@
                   width="48"
                   class="h-10 w-10 p-1 border-2 rounded
                   bg-slate-50 hover:bg-slate-300
-                  {tool['modes'].includes(selectedTool) ? 'border-main' : 'border-transparent'}"
+                  {Object.values(tool.modes).includes(selectedTool)
+                    ? 'border-main'
+                    : 'border-transparent'}"
                 >
                   <title>{tool.name}</title>
                   <path d={tool.icon} fill="currentcolor" />
@@ -56,7 +58,7 @@
                 class="absolute inset-y-0 left-16 border rounded-lg z-10 hidden group-hover:flex
                 shadow bg-slate-50 border-slate-300"
               >
-                {#each tool["modes"] as mode}
+                {#each Object.values(tool.modes) as mode}
                   <button class="w-full p-4" on:click={() => selectTool(mode)}>
                     <svg
                       xmlns="http://www.w3.org/2000/svg"