+++ ui/components/core/src/interfaces.ts
@@ -15,53 +15,121 @@
 
 // Exports
 
-export interface Dict<Type> {
-  [key: string]: Type;
-}
+// DATASET
 
-export interface ItemDetails {
-  itemData: ItemData;
-  itemObjects: Array<ObjectData>;
-}
-
-export interface ItemData {
+export interface DatasetInfo {
   id: string;
-  views: Dict<ViewData>;
-  features: DatasetItem;
+  name: string;
+  description: string;
+  estimated_size: string;
+  num_elements: number;
+  preview: string;
+  splits: Array<string>;
+  tables: Record<string, Array<DatasetTable>>;
+  categories: Array<DatasetCategory>;
+  stats: Array<DatasetStat>;
+  page?: DatasetItems;
 }
 
-export interface ViewData {
+export interface DatasetTable {
+  name: string;
+  fields: Record<string, string>;
+  source?: string;
+  type?: string;
+}
+
+export interface DatasetCategory {
+  id: number;
+  name: string;
+}
+
+export interface DatasetStat {
+  name: string;
+  type: string;
+  histogram: Array<Record<string, number | string | boolean>>;
+  range?: Array<number>;
+}
+
+// DATASET ITEM
+
+export interface DatasetItem {
   id: string;
+  split: string;
+  views: Record<string, ItemView>;
+  objects: Record<string, ItemObject>;
+  features: Record<string, ItemFeature>;
+  embeddings: Record<string, ItemEmbedding>;
+}
+
+export interface DatasetItems {
+  items: Array<DatasetItem>;
+  total: number;
+}
+
+// ITEM DATA
+
+export interface ItemView {
+  id: string;
+  type: string;
   uri: string;
-  height?: number;
-  width?: number;
+  thumbnail?: string;
+  frame_number?: number;
+  total_frames?: number;
+  features: Record<string, ItemFeature>;
 }
 
-export interface ObjectData {
+// ITEM OBJECT
+
+export interface ItemObject {
   id: string;
   item_id: string;
   source_id: string;
   view_id: string;
-  mask: MaskRLE;
-  bbox: BBoxXYWH;
-  category_id: number;
-  category_name: string;
+  mask: ItemURLE;
+  bbox: ItemBBox;
+  features: Record<string, ItemFeature>;
 }
 
-export interface CategoryData {
-  id: number;
-  name: string;
+export interface ItemURLE {
+  counts: Array<number>;
+  size: Array<number>;
 }
+
+export interface ItemBBox {
+  coords: Array<number>;
+  format: string;
+  is_normalized: boolean;
+  confidence: number;
+}
+
+// ITEM EMBEDDING
+
+export interface ItemEmbedding {
+  view_id: string;
+  data: string;
+}
+
+// ITEM FEATURE
+
+export interface ItemFeature {
+  name: string;
+  dtype: string;
+  value: number | string | boolean | DatasetStat;
+}
+
+// UI DATA
 
 export interface Mask {
   id: string;
   viewId: string;
   svg: MaskSVG;
-  rle?: MaskRLE;
+  rle?: ItemURLE;
   catId: number;
   visible: boolean;
   opacity: number;
 }
+
+export type MaskSVG = Array<string>;
 
 export interface BBox {
   id: string;
@@ -73,11 +141,13 @@ export interface BBox {
   opacity: number;
 }
 
-export type ItemLabels = Dict<SourceLabels>;
+// LABELS DATA
+
+export type ItemLabels = Record<string, SourceLabels>;
 
 export interface SourceLabels {
   id: string;
-  views: Dict<ViewLabels>;
+  views: Record<string, ViewLabels>;
   numLabels: number;
   opened: boolean;
   visible: boolean;
@@ -85,7 +155,7 @@ export interface SourceLabels {
 
 export interface ViewLabels {
   id: string;
-  categories: Dict<CategoryLabels>;
+  categories: Record<string, CategoryLabels>;
   numLabels: number;
   opened: boolean;
   visible: boolean;
@@ -94,7 +164,7 @@ export interface ViewLabels {
 export interface CategoryLabels {
   id: number;
   name: string;
-  labels: Dict<Label>;
+  labels: Record<string, Label>;
   opened: boolean;
   visible: boolean;
 }
@@ -109,48 +179,4 @@ export interface Label {
   bboxOpacity: number;
   maskOpacity: number;
   visible: boolean;
-}
-
-export interface Dataset {
-  id: string;
-  name: string;
-  description: string;
-  num_elements: number;
-  estimated_size: string;
-  preview: string;
-  categories: Array<CategoryData>;
-  page: DatasetItems;
-}
-
-export interface DatasetItems {
-  items: Array<DatasetItem>;
-  total: number;
-}
-
-export type DatasetItem = Array<DatasetItemFeature>;
-
-export interface DatasetItemFeature {
-  name: string;
-  dtype: string;
-  value: number | string | Stats;
-}
-
-export interface MaskRLE {
-  counts: Array<number>;
-  size: Array<number>;
-}
-
-export type MaskSVG = Array<string>;
-
-export interface BBoxXYWH {
-  coords: Array<number>;
-  format: string;
-  confidence: number;
-}
-
-export interface Stats {
-  name: string;
-  type: string;
-  range?: Array<number>;
-  histogram: Array<Dict<number | string | boolean>>;
 }